move predicate location
