+++ src/openprocurement/api/views/bid.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from logging import getLogger
-from openprocurement.api.models import Bid, get_now
+from openprocurement.api.models import Bid, get_now, isTender
 from openprocurement.api.utils import (
     save_tender,
     set_ownership,
@@ -17,7 +17,6 @@ from openprocurement.api.validation import (
 
 LOGGER = getLogger(__name__)
 
-from openprocurement.api.views.tender import isTender
 
 @opresource(name='Tender Bids',
             collection_path='/tenders/{tender_id}/bids',
